rmm: rename to `replicate` for consistency
